fix: 다음 페이지를 가져오는 로직을 외부로 분리
